*** empty log message ***

+++ ace/Thread_Manager.cpp
@@ -157,7 +157,7 @@ ACE_Thread_Manager::thr_self (ACE_hthread_t &self)
 
 // Initialize the synchronization variables.
 
-ACE_Thread_Manager::ACE_Thread_Manager (int)
+ACE_Thread_Manager::ACE_Thread_Manager (size_t)
   : grp_id_ (1)
 #if defined (ACE_HAS_THREADS)
     , zero_cond_ (lock_)
